feature - page d'erreur lors de la suppression d'un participant qui a…

+++ src/main/java/com/natixis/tricount/service/CreateUpdateListService.java
@@ -1,5 +1,6 @@
 package com.natixis.tricount.service;
 
+import com.natixis.tricount.entity.Expense;
 import com.natixis.tricount.entity.ExpenseList;
 import com.natixis.tricount.entity.Participant;
 import com.natixis.tricount.repository.ExpenseListRepository;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 @Service
 public class CreateUpdateListService {
@@ -33,9 +35,17 @@ public class CreateUpdateListService {
         participantRepository.save(participant);
     }
 
-    public void removeParticipantList(Long idParticipant)
+    public int removeParticipantList(Long idParticipant)
     {
-        participantRepository.deleteById(idParticipant);
+        Set<Expense> expenseList = participantRepository.getById(idParticipant).getExpense();
+        List<Expense> expenseListOfPayer = participantRepository.getById(idParticipant).getExpensesOfPayer();
+
+        if (expenseList.isEmpty() && expenseListOfPayer.isEmpty())
+        {
+            participantRepository.deleteById(idParticipant);
+            return 0;
+        }
+        return 1;
     }
     
 }
